+++ src/pages/infobox.jsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import { useState } from "react";
 import { deleteRecord } from "thin-backend";
 
-export default function InfoBox({time, id, openMainEditor})
+export default function InfoBox({time, id, setEditor})
 {
     const [confirm, setConfirm] = useState(false)
     
@@ -11,13 +11,18 @@ export default function InfoBox({time, id, openMainEditor})
         if (!confirm)
             setConfirm(true)
         else
-            deleteRecord('locations', id)
+            if (id != '' || id != 'new')
+            {
+                deleteRecord('locations', id)
+                setEditor("")
+            }
+    }
+    function openMainEditor()
+    {
+        setEditor("")
     }
     return (
         <div className="mini-editor">
-            <p className="time">
-                {time}
-            </p>
             <svg className={confirm ? "pencil invisible" : "pencil"} width="20" height="20" viewBox="0 0 55 55" 
             fill="none" xmlns="http://www.w3.org/2000/svg" onClick={openMainEditor}>
                 <path d="M38.9617 2.07324L33.7624 7.27246L47.7273 21.2373L52.9265 16.0381C55.6121 13.3525 55.6121 9.00195 52.9265 6.31641L48.6941 2.07324C46.0085 -0.612305 41.6579 -0.612305 38.9724 2.07324H38.9617ZM31.3347 9.7002L6.29467 34.751C5.17748 35.8682 4.36107 37.2539 3.9099 38.7686L0.107166 51.6914C-0.161389 52.6045 0.0856813 53.582 0.751697 54.248C1.41771 54.9141 2.39525 55.1611 3.2976 54.9033L16.2204 51.1006C17.7351 50.6494 19.1208 49.833 20.238 48.7158L45.2996 23.665L31.3347 9.7002Z" fill="black"/>
