+++ src/styles/minieditor.css
@@ -1,10 +1,16 @@
 .mini-editor {
     display: flex;
     background-color: #d9d9d9;
-    padding: 0.5em;
-    border-radius: 0.5em;
-    align-items: center;
-    position: relative;
+    padding: 1em;
+    border-radius: 0px 10px 0px 0px;
+    box-shadow: 0px -5px 5px rgba(0, 0, 0, 0.2);
+
+    justify-content: space-around;
+    width: 20vw;
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    z-index: 999;
 }
 
 .mini-editor svg {
